fix test

+++ tests/test_orientation_estimation/test_orientation_estimation.py
@@ -34,7 +34,7 @@ class TestGyroIntegration:
         fs = 100
         window_width = 8
         # start at window_with/2 because GyroIntegration._calculate_initial_orientation uses start+-half window size
-        start_sample = np.floor(window_width/2)
+        start_sample = int(np.floor(window_width/2))
         # 180 degree rotation around i_axis
         sensor_data, event_list = self.get_dummy_data(start_sample, axis_to_rotate, fs)
